+++ src/components/Providers/UserProvider.js
@@ -1,5 +1,5 @@
 import React, { createContext, useContext, useState, useEffect } from "react";
-import instance from "api/axios";
+import instance from "queries/axiosInstance";
 import handleAPIError from "utils/handleAPIError";
 
 const UserDataContext = createContext({
